Move the folder details outside folder collection in the store
