+++ src/core/main/video-detail.js
@@ -3,7 +3,7 @@ import { useParams } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 //own
 import Modal from "components/modal";
-import { detailActions } from "store/collections/details/detail-slice";
+import { detailActions } from "store/details/detail-slice";
 
 const VideoDetail = () => {
   const dispatch = useDispatch();
