+++ src/store/details/detail-slice.js
