+++ src/store/details/get-detail.js
@@ -5,7 +5,7 @@ import {
   URL_VIDEO_MOVIES,
   URL_TV_VIDEO,
 } from "constants";
-import { detailActions } from "store/collections/details/detail-slice";
+import { detailActions } from "store/details/detail-slice";
 
 export const getDetailFilm = (id, category) => {
   return async (dispatch) => {
